Checkout Controller checkoutPayment done.

+++ database/migrations/2021_05_07_123845_create_order_items_table.php
@@ -16,7 +16,7 @@ class CreateOrderItemsTable extends Migration
         Schema::create('order_items', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->unsignedInteger('quantity');
-            $table->decimal('price', 20, 6);
+            $table->float('price');
 
             $table->foreignId('order_id')->constrained('orders')->onDelete('cascade');
             $table->foreignId('product_id')->constrained('products')->onDelete('cascade');
